+++ src/lib.rs
@@ -1,0 +1,3 @@
+// Exports which are used in integration/e2e test cases.
+pub mod device;
+pub mod util;